+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "The optimal urinal problem".
 - Tests for "Light bulbs".
 - Tests for "What the brainfuck !".
+- Tests for "Bouncing barry".
 
 ## [1.11.0] - 2022-07-31
 ### Added